moved queue functions inline to judging.ts
